requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pymysql==1.1.1
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
neo4j==5.24.0
openai==1.52.2
